Refactoring. Incapsulate internal logic in single class and expose extension point for extension attributes.
