--- InventoryInStorePickupGraphQl/Model/Resolver/PickupLocations/SearchRequest/CurrentPage.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest;
-
-use Magento\InventoryInStorePickupApi\Model\SearchRequestBuilderInterface;
-
-/**
- * Resolve Current Page parameters.
- */
-class CurrentPage implements ResolverInterface
-{
-    /**
-     * @inheritdoc
-     */
-    public function resolve(
-        SearchRequestBuilderInterface $searchRequestBuilder,
-        string $fieldName,
-        string $argumentName,
-        array $argument
-    ): SearchRequestBuilderInterface {
-        $searchRequestBuilder->setCurrentPage($argument[$argumentName]);
-
-        return $searchRequestBuilder;
-    }
-}
