+++ InventoryInStorePickupGraphQl/Test/GraphQl/PickupLocationsPositiveCasesTest.php
@@ -32,16 +32,16 @@ class PickupLocationsPositiveCasesTest extends GraphQlAbstract
 
     /**
      *
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/products.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/sources.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryInStorePickupApi/Test/_files/source_addresses.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryInStorePickupApi/Test/_files/source_pickup_location_attributes.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/stocks.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/stock_source_links.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventorySalesApi/Test/_files/websites_with_stores.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventorySalesApi/Test/_files/stock_website_sales_channels.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryInStorePickupApi/Test/_files/inventory_geoname.php
-     * @magentoApiDataFixture ../../../../app/code/Magento/InventoryApi/Test/_files/source_items.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryApi/Test/_files/products.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryApi/Test/_files/sources.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryInStorePickupApi/Test/_files/source_addresses.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryInStorePickupApi/Test/_files/source_pickup_location_attributes.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryApi/Test/_files/stocks.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryApi/Test/_files/stock_source_links.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventorySalesApi/Test/_files/websites_with_stores.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventorySalesApi/Test/_files/stock_website_sales_channels.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryInStorePickupApi/Test/_files/inventory_geoname.php
+     * @magentoApiDataFixture ../../../../ext/magento/inventory/InventoryApi/Test/_files/source_items.php
      *
      * @magentoConfigFixture cataloginventory/source_selection_distance_based/provider offline
      *
