--- InventoryInStorePickupGraphQl/etc/graphql/di.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest">
-        <arguments>
-            <argument name="resolvers" xsi:type="array">
-                <item name="area" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Area</item>
-                <item name="filters" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Filter</item>
-                <item name="currentPage" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\CurrentPage</item>
-                <item name="pageSize" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\PageSize</item>
-                <item name="sort" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Sort</item>
-                <item name="productsInfo" xsi:type="object">Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\ProductsInfo</item>
-            </argument>
-        </arguments>
-    </type>
-</config>
